add null check, start test split
